sedimcorp_certificados/requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.5
colorama==0.4.6
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mysqlclient==2.2.8
pillow==12.1.1
PyJWT==2.11.0
python-dotenv==1.2.2
qrcode==8.2
reportlab==4.4.10
sqlparse==0.5.5
tzdata==2025.3
